Promoted Deployment and Results data classes.
