+++ workers/unity/Assets/Fps/Scripts/UI/Controllers/ResultsScreenController.cs
@@ -95,23 +95,5 @@ namespace Fps
             entry.SetAllTextVisuals(isLocalPlayer ? LocalPlayerTextColor : DefaultTextColor, isLocalPlayer);
             entry.SetBackgroundColor(isLocalPlayer ? LocalPlayerBackgroundColor : currentBgColor);
         }
-
-        public class ResultsData
-        {
-            public readonly int Rank;
-            public readonly string PlayerName;
-            public readonly int Kills;
-            public readonly int Deaths;
-            public readonly float KillDeathRatio;
-
-            public ResultsData(int rank, string playerName, int kills, int deaths)
-            {
-                Rank = rank;
-                PlayerName = playerName;
-                Kills = kills;
-                Deaths = deaths;
-                KillDeathRatio = deaths == 0 ? kills : kills / (float) deaths;
-            }
-        }
     }
 }
