+++ workers/unity/Assets/Fps/Scripts/UI/DeploymentData.cs
@@ -1,0 +1,18 @@
+﻿namespace Fps
+{
+    public class DeploymentData
+    {
+        public readonly string Name;
+        public readonly int CurrentPlayers;
+        public readonly int MaxPlayers;
+        public readonly bool IsAvailable;
+
+        public DeploymentData(string name, int currentPlayers, int maxPlayers, bool isAvailable)
+        {
+            Name = name;
+            CurrentPlayers = currentPlayers;
+            MaxPlayers = maxPlayers;
+            IsAvailable = isAvailable;
+        }
+    }
+}
