+++ workers/unity/Assets/Fps/Scripts/UI/DeploymentData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5c0ca622af632c741b79376908ddea04
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
