+++ workers/unity/Assets/Fps/Scripts/UI/ResultsData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 92a67a8de5851f549a1208242ea69f2a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
